+++ app/components/ExploreAccordionSections.jsx
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import exploreStyles from '../styles/explore.css';
-import {Button, Card, CardTitle, CardSubtitle, CardHeader, CardText, Fab} from 'react-mdc-web';
+import {Button, Card, CardTitle, CardSubtitle, CardHeader, CardText, Fab, Icon} from 'react-mdc-web';
 import {getColor} from '../utils/getConfig';
 import ol from 'openlayers';
 
@@ -8,7 +8,15 @@ import ol from 'openlayers';
 class ExploreAccordionSections extends Component {
     constructor(props) {
         super(props);
+        this.state = {
+            inner_accordion_icon: false,
+        };
         (this:any).clickSensor = this.clickSensor.bind(this);
+        (this:any).clickedInnerAccordion = this.clickedInnerAccordion.bind(this);
+    }
+
+    clickedInnerAccordion() {
+        this.setState({inner_accordion_icon: !this.state.inner_accordion_icon});
     }
 
     clickSensor(id, name, coordinates, e) {
@@ -26,7 +34,13 @@ class ExploreAccordionSections extends Component {
 
             let lonLatPoint = [data.geometry.coordinates[0], data.geometry.coordinates[1]];
             let webMercatorPoint = ol.proj.fromLonLat(lonLatPoint);
+
             let color = getColor(data.properties.type.id);
+            if (this.props.selectedSensorID !== data.id) {
+                // Add opacity to the HEX color (50%)
+                color = color + '80';
+            }
+
             let longer_name = '';
             if (data.name.length >= 6) {
                 longer_name = '...';
@@ -36,31 +50,46 @@ class ExploreAccordionSections extends Component {
                 help_text += " - " + data.properties.popupContent;
             }
             let button_label = data.properties.name.substring(0, 6).trim().replace("-", "_").replace(" ", "_");
-
-            item_pills.push(
-                <Button key={data.id} className={exploreStyles.exploreButton}
-                        style={{backgroundColor: color}} id={data.id} title={help_text}
-                        onClick={() => {this.clickSensor(data.id, data.name, webMercatorPoint)}}>
+            if (this.state.inner_accordion_icon) {
+                item_pills.push(
+                    <Button key={data.id} className={exploreStyles.exploreButton}
+                            style={{backgroundColor: color}} id={data.id} title={help_text}
+                            onClick={() => {
+                                this.clickSensor(data.id, data.name, webMercatorPoint)
+                            }}>
                             <span className={exploreStyles.exploreButtonText}>
                                 {button_label}{longer_name}
                             </span>
-                </Button>
-            )
+                    </Button>
+                )
+            } else {
+                item_pills = '';
+            }
 
         });
 
-        if (item_pills.length > 0) {
+            let color = getColor(this.props.sourceId);
+            // Add opacity to the HEX color (75%)
+            color = color + 'BF';
+
             contents.push(
                 <div key={this.props.id}>
-                    <div>
-                        <span data-tooltip={this.props.tooltipVal} className={exploreStyles.regionLabel}>
-                            {this.props.sectionLabel} ({item_pills.length})
-                        </span>
+                    <div onClick={() => {this.clickedInnerAccordion()}}>
+                        <div style={{backgroundColor: color}}>
+                            <span data-tooltip={this.props.tooltipVal} className={exploreStyles.regionLabel}>
+                                {this.props.sectionLabel}
+                            </span>
+                            <span className={exploreStyles.regionCount}>
+                                ({source_data.length})
+                                <Icon className={"material-icons " + exploreStyles.accordionIcon}
+                                      name={this.state.inner_accordion_icon ? 'expand_less' : 'expand_more'}
+                                />
+                            </span>
+                        </div>
                     </div>
                     <div className={exploreStyles.sectionPills}>{item_pills}</div>
                 </div>
-            )
-        }
+            );
 
         return (
             <div>{contents}</div>